adjusted schema and coresponding files to use truck database by truck owners to update data on trucks.
