+++ controllers/truckController.js
@@ -8,6 +8,25 @@ module.exports = {
             .sort({ date: -1 })
             .then(data => res.json(data))
             .catch(err => res.status(422).json(err));
+    },
+    create: function (req, res) {
+        db.Truck
+            .create(req.body)
+            .then(data => res.json(data))
+            .catch(err => res.status(422).json(err));
+    },
+    update: function (req, res) {
+        db.Truck
+            .findOneAndUpdate({ _id: req.params.id }, req.body)
+            .then(data => res.json(data))
+            .catch(err => res.status(422).json(err));
+    },
+    remove: function (req, res) {
+        db.Truck
+            .findById({ _id: req.params.id })
+            .then(data => data.remove())
+            .then(data => res.json(data))
+            .catch(err => res.status(422).json(err));
     }
 
 };
