+++ models/index.js
@@ -1,5 +1,4 @@
 module.exports = {
   Account: require("./account"),
   Truck: require("./truck"),
-  Owner: require("./truckOwner"),
 };
