--- routes/api/owners.js
@@ -1,17 +1,0 @@
-const router = require("express").Router();
-const truckOwnersController = require("../../controllers/truckOwnersController");
-
-
-// Matches with "/api/owners
-router.route("/")
-    .get(truckOwnersController.findAll)
-    .post(truckOwnersController.create);
-
-// Matches with "/api/owners/:id"
-router
-    .route("/:id")
-    .get(truckOwnersController.findById)
-    .put(truckOwnersController.update)
-    .delete(truckOwnersController.remove);
-
-module.exports = router;
